+++ pkg/backend/httpstate/client/client.go
@@ -97,8 +97,20 @@ func getStackPath(stack StackIdentifier, components ...string) string {
 	return path.Join(append([]string{prefix}, components...)...)
 }
 
-// publishPolicyPackPath returns the API path to for the given organization with the given
-// components joined with path separators and appended to the organization root.
+// listPolicyGroupsPath returns the path for an API call to the Pulumi service to list the Policy Groups
+// in a Pulumi organization.
+func listPolicyGroupsPath(orgName string) string {
+	return fmt.Sprintf("/api/orgs/%s/policygroups", orgName)
+}
+
+// listPolicyPacksPath returns the path for an API call to the Pulumi service to list the Policy Packs
+// in a Pulumi organization.
+func listPolicyPacksPath(orgName string) string {
+	return fmt.Sprintf("/api/orgs/%s/policypacks", orgName)
+}
+
+// publishPolicyPackPath returns the path for an API call to the Pulumi service to publish a new Policy Pack
+// in a Pulumi organization.
 func publishPolicyPackPath(orgName string) string {
 	return fmt.Sprintf("/api/orgs/%s/policypacks", orgName)
 }
@@ -495,6 +507,26 @@ func (pc *Client) StartUpdate(ctx context.Context, update UpdateIdentifier,
 	}
 
 	return resp.Version, resp.Token, nil
+}
+
+// ListPolicyGroups lists all `PolicyGroups` the organization has in the Pulumi service.
+func (pc *Client) ListPolicyGroups(ctx context.Context, orgName string) (apitype.ListPolicyGroupsResponse, error) {
+	var resp apitype.ListPolicyGroupsResponse
+	err := pc.restCall(ctx, "GET", listPolicyGroupsPath(orgName), nil, nil, &resp)
+	if err != nil {
+		return resp, errors.Wrapf(err, "List Policy Groups failed")
+	}
+	return resp, nil
+}
+
+// ListPolicyPacks lists all `PolicyPack` the organization has in the Pulumi service.
+func (pc *Client) ListPolicyPacks(ctx context.Context, orgName string) (apitype.ListPolicyPacksResponse, error) {
+	var resp apitype.ListPolicyPacksResponse
+	err := pc.restCall(ctx, "GET", listPolicyPacksPath(orgName), nil, nil, &resp)
+	if err != nil {
+		return resp, errors.Wrapf(err, "List Policy Packs failed")
+	}
+	return resp, nil
 }
 
 // PublishPolicyPack publishes a `PolicyPack` to the Pulumi service. If it's successful, it returns